make example code a working example
Add verify callback to client such that it accepts all certificates.
Include boost beast example certificates.
Note that the actual certificates do not matter. The client will accept any certificate, the server will reject
any certificate. It is only such that the handshake does not fail because of missing certificates.